fix: inserting remote elements shared by other sequences
When an Element is inserted into a Sequence, it must not be
pointed to by any other Sequence, because each Sequence will attempt to
free every Element that it contains.

If an Element is shared by two Sequences, a double-free error may occur.

Therefore, implement a deep copy function to create a copy of an Element
such that there are no shared references between them.